Refactor: Move WalkForwardValidator to its own module
This commit reorganizes the code by moving the WalkForwardValidator class and its associated functions from `src/validation/walk_forward_validator.py` to `src/training/steps/backtesting/walk_forward_validation.py`. This improves modularity and organization.
